fix(js): integrate handleError in NodeJSAnoncreds
